feat(issue-167): consolidate quality test suites
